+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: add execution listener headers templating support ([#233](https://github.com/camunda/element-templates-json-schema/pull/233))
 * `FIX`: deprecate `optional` property for input mappings ([#235](https://github.com/camunda/element-templates-json-schema/pull/235))
 
 ## 0.39.2